refactor(client): extract the Memories tab into a self-loading OnPush component (A17.9b-6d)
First record tab out of the shell. memories-tab.component owns the memory create
form, the inline edit, delete, and its own search / type-tag filter / pagination +
list loader. It reads records + derived views from BrainStore, shares the singleton
load/edit/delete interaction with the shell via RecordListState, uses EntityRefPicker
for entity chips and RecordDrawerState to open the detail drawer.

Self-loading: the shell renders it behind `@if (activeTab() === 'memories')`, so it is
created on activation and destroyed on switch. An effect on the spaceId input loads on
creation and reloads on a space switch while mounted; the shell's loadCurrentTab now
early-returns for memories (it no longer drives the tab's loading). Create/delete emit
a `mutated` output so the shell refreshes the tab-count stats — the one legitimate
output (tab counts are parent view-state).

The brain-scoped record-table CSS (search header, create form, filter chips, inline
confirm, description cell) moved to a shared brain-table.styles.ts the tab components
import; the table/pagination/empty-state/tag styles are global. The shell's inline
copies stay until the last record tab leaves, then a cleanup removes the dead CSS.

Tests: the 6b memories characterization cases + the two memories-table rendering tests
relocated to memories-tab.component.spec.ts, plus new self-load and `mutated`
assertions. brain.component.ts 2416 -> 2070. Client suite 188 -> 190. tsc
--noUnusedLocals clean; prod AOT build warning-free bar the pre-existing qrcode bailout.
